Implemented Step In
First I needed to make sure Page2 and Page2 opcodes are handled as extensions to the following opcode so we don't try to single step past the Page2 or Page3 separately. I did this by changing the instruction counter to be at the end of the loop and then doing a continue when Page2 or Page3 are encountered so it never tries to return after executing these. Then I put the test for Step In at the end of the loop so it's also skipped.

Works on both Mac and ESP for simple.asm