Got rid of page2 and page3 entries in opcodeTable.
For a few opcodes the page2 and page3 variants just use different registers. For these cases I added some pseudo register values in the enum (e.g., XYS). When in the main page it selects the first register (e.g., X) and for the other pages it selects the two alternates. This gets rid of 8 bits in the opcodeTable entry, saving 256 bytes in the engine.